requirements.txt
ImageHash==4.1.0
Pillow==7.2.0
decord==0.4.0
tqdm==4.48.2
